Convertiti valori da passare alle tabelle/grafici in double per problemi di precisione (ticket #2788)

+++ energy_graph.h
@@ -50,7 +50,7 @@ public:
 
 public slots:
 	void init(int bars, QString t, const QMap<int, QString> &x_axis);
-	void setData(const QMap<int, float> &data);
+	void setData(const QMap<int, double> &data);
 
 protected:
 	void paintEvent(QPaintEvent *e);
@@ -76,7 +76,7 @@ private:
 #ifdef TEST_ENERGY_GRAPH
 public:
 #endif
-	QMap<int, float> graph_data;
+	QMap<int, double> graph_data;
 };
 
 
@@ -91,7 +91,7 @@ Q_OBJECT
 public:
 	EnergyTableContent(int n_dec);
 	void init(int num_val, QString left_text, QString right_text, int shift_val=-1);
-	void setData(const QMap<int, float> &data);
+	void setData(const QMap<int, double> &data);
 	void setNumDecimal(int d);
 
 public slots:
@@ -116,7 +116,7 @@ private:
 
 private:
 	QString border_color, heading_color, odd_row_color, even_row_color, text_color;
-	QMap<int, float> table_data;
+	QMap<int, double> table_data;
 	QString left_text, right_text;
 	int num_values;
 	int shift_value;
@@ -132,7 +132,7 @@ Q_OBJECT
 public:
 	EnergyTable(int n_dec);
 	void init(int num_val, QString left_text, QString right_text, QString date, int shift_val=-1);
-	void setData(const QMap<int, float> &data);
+	void setData(const QMap<int, double> &data);
 	void setNumDecimal(int d);
 
 private:
